Added VatRate, Countries and Sales Invoices API methods.

+++ Validators/BankAccountValidator.cs
@@ -4,7 +4,7 @@ namespace ERPXTpl.Validators
 {
     internal class BankAccountValidator
     {
-        internal static string GetBankAccountById(int bankAccountId)
+        internal static string GetBankAccountById(long bankAccountId)
         {
             if (bankAccountId == 0)
             {
